Create README.md
Respected Sir,
I have made a program on ATM service.
In the program I have used python function def, while loop and if, elif, else statement.
In the program, the user has to provide its valid name of their bank account and the balance of the account.
The machine will then show the menu which will give the user the option of check balance, withdraw, deposit or quit the service.Each option has its own function to be carried out. When balance function is called it will show the amount present in the account. The withdraw function, when called will ask the user the amount they want to withdraw and then subtract that amount from the account, thereby changing the balance of the account. The deposit function will ask the user the amount they want to deposit and will add the amount to the account, thereby changing the balance of the account.The quit function will end the service.